move secrets getting and db connection to common layer to reduce duplication
